Refactor the list interface MAE wrapper.
