+++ gridworld/__init__.py
@@ -1,4 +1,4 @@
 __version__ = "0.0.1"
 from .base import ComponentEnv, MultiComponentEnv
 from .multiagent_env import MultiAgentEnv
-from .multiagent_env_openai_style import MultiagentEnvOpenAIStyle
+from .multiagent_list_interface_env import MultiAgentListInterfaceEnv
